refactor: query param serviceStatus
