--- frontend/constants/components/serviceStatus.ts
@@ -1,1 +1,0 @@
-export const SERVICE_STATUSES = ['보충역', '현역'];
